Fix bug in `heap`
The `split` function had a bug in that it tried to access the `next` field of the header even if it was a `nullptr`. This caused sort of undefined behaviour and was over-all not good. With the new fix, I no longer get page faults/general protection faults when malloc is called.