Beginnings of FatFS
I figured the first drive I should be able to read is probably the boot drive... So why not write a driver for FAT? It's one of the simplest file systems, so it's not bad to start off with. I realize I'm missing a VFS at the moment but everything is in the very early stages. Once I get a filesystem or two up and running, it will be easier to know what I will need in a VFS.