[Heap]: Implement operate delete with two arguments
This has been in the standard since C++14 and often my compiler will generate calls to this delete instead of the normal one for seemingly no reason (often has to do with pointer arrays). The standard most of the time just falls through to the regular delete with one argument (void*), so that's exactly what I did.