Write RTC time to screen, serial, among other things
The Real Time Clock can now be updated at any time from within the kernel using the `get_date_time` function (previously private to the RTC class). This is most probably temporary, but allows for a pretty ticking clock to be printed in the top right of the kernel, for now.

NOTE: "among other things" refers to the following:
- Slight switch-up in logic of temporary `print_now` function in  `kernel.cpp`
  - No longer swap buffers at end of function.
  - Take in offset parameter to pass to newline.
- Clear screen after kernel_init has returned (cleaner)
  - Blue-Red face is now only seen at boot-time.
- Todo comments, as usual.
- Some more debug information during `kernel_init`.