Entire re-design of FAT file system
The structure is now less of a garbage fire, although it still has it's quirks.

I've removed the old system, as the new one has equal if not more functionality already.

The VFS is entirely un-implemented still, but the beginnings of how it works are starting to take shape in my head.