Port shell scripts from `sh` to `bash`
This allows the script to get it's own directory, allowing them to be run from anywhere yet still work. This greatly enhances the workflow as after building in the cmake directory, the run script can be run without any `cd`ing required.

This does break the scripts if symlinks are involved in the path, but in our case (and the common case when developing for LensorOS), this doesn't matter.