Update README: Info concerning serial output
On Windows, by doing the intuitive thing and running the `run.bat` by double-clicking it from Windows Explorer, `cmd` is opened; this is a problem as it is not a terminal that supports ANSI color codes (at least on my machine although people online seem to have different results). To solve this, there is a better Windows Terminal developed open-source by Microsoft themselves that I highly recommend using for osdev on a Windows machine.