[UART]: Add ability to hide serial output color codes (SGR)
This allows for people with terminals that don't support ANSI color codes to still be able to read the serial output (otherwise everything is mangled with escape characters and ANSI codes).

Simply define `LENSOR_OS_UART_HIDE_COLOR_CODES` somewhere in the project, or when compiling. This causes the UART `writestr` function to not print any recognized ANSI color codes, SGRs, CSIs, etc.

This is not very clean code, definitely not bug-free, but it should get the job done so that somebody who needs this feature can enable it and actually be able to read the serial output on startup. Ideally I will come back and fix this at a later date (as with everything in this project).