[Rand]: Add simple random number generator
It's only a linear congruential generator, so the distribution is borderline unusable, but it's okay to use for silly things that won't be around forever (like mouse draw color).

The only reason I decided to create a whole class is because I presume getting a quick random number will be something I need to do a lot for testing, but the randomness of that number isn't exactly at question here.