[UART]: Remove unused
The MAX_SERIAL_STRING_LENGTH definition was good in theory but I don't think it's required as no external programs will be using the internal UART driver. As all input can be trusted (as it is from our own kernel), there's no need for a sanity check like that one.

Also I moved the declaration of the global serial driver to the top of `uart.cpp`. I feel like having the `extern` at the bottom of the header and the declaration at the top of the `.cpp` is better for readability, and it makes it so I have a nice simple rule for where to place those things.