[TSS]: Finalize setup, create test user function (it works!!!)
For test purposes, all memory is user accessible and page tables are not swapped when switching to user mode. This is about as bad as it gets security wise and really just in general.

Slowly, however, it will improve.

For now, I am happy to be able to enter a function in userland/user mode/ring3/whatever you want to call it, and create system calls to the kernel for specific, outlined behaviour. I can't wait to build on this system even more and more and more (I really need to sketch out a GUI system and how that might work).