[UART]: Basic serial input
The characters you type into the serial input are now output using the graphical renderer to the screen! This feels like a large step for the OS, as it could now technically be controlled from another computer (which is kind of cool).

Obviously, this is temporary (don't want serial input to draw to the screen forever), but for now it's a good way to visualize what's going on.

- Installed IRQ interrupt handler for UART seral communications chip: port=`COM1`.
- Enabled `Data Recieved` interrupts in UART.
- All keyboard functionality has been moved inside the `Keyboard` namespace.
- Sneak a comment in `pci.cpp`.
- Make UART code more readable using pre-processor definitions