Keyboard IO: Completely refactor text renderer
This new renderer adds a text cursor that is able to be positioned using the arrow keys. This allows for much more control over the text editing (like vim's default mode, effectively).

This system is temporary, and far from perfect/even okay, but it will get the job done for now.