Successfully run with `-O4` compiler flag
This is the highest level of compiler optimization applied to the code, so the fact that it works implies that the compiler understands what I am telling it I want to do (if that makes sense).

I believe keeping heavy optimizations on by default will encourage best practices in code, and keep the system running nice and smooth (although it is fast without it, as well).

Errors at compile time: ✔
Errors at run time: 🚫