[AHCI]: Use Port's configured buffer by default.
It's slightly hack-ish with the default being nullptr, and within the function the address is checked against nullptr; if it matches, it is set to the AHCI Port's configured buffer. This works just fine, but might have unforeseen consequences at some point in the future (but really it should be fine). This is done so as to not require passing a buffer to the `Read()` function, as it causes confusion (address must be physical, no heap buffers allowed). To extract data into a heap-allocated buffer, one must `memcpy()` from the port's buffer to the heap-allocated one.