OVMF: Include smaller binaries
The default image size is 4MiB, which is quite large, in my opinion. By building it with `FD_SIZE_IN_KB=2048`, the generated image is half the size. This will speed up downloads, and really has no downside, at least for now. We can always move back to 4MiB images if it becomes an issue.

Simplify OVMF build instructions: we no longer are required to go back and forth editing a text file each time a different package needs to be built; instead use the command line interface to specify flags and options.