OVMF: Use a copy of OVMF_VARS
This allows it to never show up in the working tree of git, without requiring any workarounds like `--assume-unchanged` (or actual solutions like running `--skip-worktree`).

This way, the modified files won't get tracked (as they don't need to be when modified by QEMU for a local session), but any updates to the default will be updated and used only if the old one is deleted.

Also just going to mention how crazy the if/else syntax is in bash.