HPET: read/write now uses volatile read/write
HPET `readl()` and `writel()` functions now use a volatile read and volatile write respectively. This allows for the compiler to know not to optimize these out, and that the values at these addresses may change on a whim.

I can't believe I tried to set a bit to a `1` with an `&`... I'm disappointed.