HPET: Clean-up, Refactor ACPI prep
The ACPI preparation function with `kUtility` kind of started to handle everything, so I decided to move ACPI initialization to the ACPI namespace and file.

As for the growing responsibilities of the old prep function, they were handled as follows:
- PCI Preparation moved to `prepare_pci()` function (most likely will move to `PCI::initialize()` at some point).
- HPET Timer initialization now handles finding table within ACPI.

To make all this work, it is necessary to `prepare` the ACPI by finding the eXtended System Descriptor Table (XSDT) and caching it (`ACPI::gXSDT`). Then, a function that finds an ACPI table by signature alone can use the cached XSDT to find tables within.