PIT: Auto-initialize on construction
This did require an overhaul of the scheduler; it is now a class instead of a namespace. This allows certain classes to `friend` the scheduler which is very useful. The members of the class are all static, meaning the usage is not much different than before with the namespace setup.

Move PIT `Ticks` member to private, as only itself should be able to modify it (or the interrupt handler which calls `tick()`).

PIT include guard in the header still had the old `timer` name, not `PIT`, as it should be.