IRQ0: A better default
If the scheduler is not used (for some reason that may appear in the future), then the interrupt handler for IRQ0 installed in the IDT mustn't be the scheduler specific one. To accomplish this, the scheduler installs it's handler into the IDT, and by default the IDT stores a very simple handler that increments the clock.