Heap: The bug is fixed :^)
Turns out you can't use a pointer iterator when it's a linked list (duh!)... Fixed it though, and it seems to work now.

While writing the above paragraph, I realized the reason `heap_print_debug()` wasn't showing the last region of the heap (the free one, usually): the `do{ ... }/while()` loop was comparing against the iterator's next, but that was only happening after the iterator was assigned to the "current" iteration's next, due to previous bugfix.