Move UARTDriver to a namespace
The whole thing about keeping it on the heap was originally just a way to test the heap in a real way. Now I understand all I need is a namespace and to call an initialization function once before using it.

The following shell command helped a huge amount in the search and replace function:
```shell
grep -rl "findstr" path | xargs sed -i "s/findstr/replacestr/g"
```