Serial output formatting
Also try to move some debug output from `kUtility` into the places where they should actually be used (as the `srl` UARTDriver was a pointer before, it was required to be instantiated before use, causing a few things to have to wait to be output to the debug output. The previous simplification of the serial output allows me not to worry about whether or not the driver is created before trying to use it (as if it is not initialized it will do nothing).