Update programmable interrupt timer interface
This was one of the first things I ever wrote, and it was quite awful.

Enums are now taken advantage of to more clearly represent what is happening in the code. Channels can now be configured with a helper function and no longer require manual IO, etc.

Beginnings of a PC Speaker implementation