Enable SSE, XSAVE, AVX
All is going well. There is also a difference here between VBOX and QEMU; VBOX supports AVX and XSAVE, while QEMU does not! This makes sense as QEMU reports a vendor id of "AuthenticAMD" (which I find amusing), and VBOX reports a vendor id of "GenuineIntel" (equally so), and AVX was introduced by Intel in 2011.