Move scripts to `/kernel/scripts/` directory
This cleans up the kernel directory a lot, as it was getting bloated with scripts (and still is).

The run scripts will be moved next, but I found one interesting thing while doing this: bash scripts can be run on Windows. The `run.sh` actually works just fine on Windows, but there's a kicker: the terminal doesn't detect colorful output from bash, so all the color codes mangle the output. This can of course be solved by LensorOS internally using the corresponding preprocessor directive, but I prefer nice, colorful output (so the `.bat` will have to stay, but maybe just as a wrapper around the `.sh` instead of two copies of the same thing).