Build for VBOX by default
This allows for the bootable disk image to still be created (allowing QEMU to boot from it), and also allows for the bootable el torito CD-ROM image to be created and used by VBOX to boot from without crashing. This is only necessary because parts of LensorOS are not yet fully functional in VBOX (HPET initialization causes immediate crash in VBOX).