Begin `libc` implementation
It's very messy. I'm sure it's full of bugs, but this is the minimum required to get libgcc to build with the sysroot set and everything (at least on my machine, right now).

I'm not proud of a lot of the code here, and I'm even less proud of the documentation, but I hope it's enough to get things started.