Linked List: Implement `at(index)` API
As for error handling, it currently just returns the last non-null node's value in the list if given an index larger than is actually available, which is questionable behaviour but I don't want to overhaul the linked list right now with error propagation... For that reason there is a large FIXME slapped in the function that says just that, and outlines why the current way is bad.

Also don't jump to "userland" anymore in the main `kernel.cpp` file.