Linked List: Impl. array subscript operator
This required a slight reworking of how the SinglyLinkedListNode::value() functions work, but in the end it is definitely for the better.

I did have to remove a null dereference check, which means if `at()` or the `[]` operator are used on an empty list the kernel will null-dereference (BAD!!!). I need an `ErrorOr<T>` or something that can propagate errors or a valid result (almost like `Optional`.