Very minor clean up that ultimately does nothing
I spent half an hour trying to store the Port pointers in a linked list but it caused the last point on the heap point to the third last, resulting in an infinite loop when traversing the heap. I don't know how this happens, at all, and I hate that I'm too stupid to figure this out. Hopefully someone less stupid is making the OS of tomorrow, and I will be forgotten. I don't deserve anything, and I want nothing.