Begin the boot info copying kerfuffle
The BootInfo structure is full of pointers to structures that also need to be copied to the new address space.

I presume once all of memory isn't identity mapped in the kernel anymore, the basic renderer and anything else using hardware memory, will need to map the framebuffer into virtual memory on their own.