Further and further on, the trot keeps up
It's a slow pace, but it's nice to be making progress. We are so close to being back at 100% functionality while also executing the kernel from higher half. This has been an insane transition and I'm happy that it has happened so quickly.

The main issue right now is that the prekernel page map is very barebones, and the physical memory manager relies on finding a portion of memory from the EFI memory map to store the initial physical page bitmap in; this portion is not mapped in the prekernel page map, so there is a lot of chicken/egg dilemma going on here.

To fix, I could either:
- Use an initial bitmap that corresponds to a fixed amount of allocated memory (max amount of space hard-coded), but I'm not sure how to use the full amount of RAM on the system this way.
- Map the range virtually (difficult to do while setting up the physical memory manager...)