Renderer: If no room for deferred FB, don't use one
I am chasing down problems with VBOX booting, as well as the issue of the FAT driver now just crashing immediately no matter what (may be due to me never copying the ACPI RSDP2 in the prekernel). Either way, the next commits may seem unrelated to a higher half kernel at first glance.

Basically, if the memory allocation for the second framebuffer fails (likely due to not enough memory or bugs in the memory management), just assign the Target pointer to the same as the Render (effectively only using one buffer).

To combat the excessive amount of `memcpy()`s that would occur trying to copy itself to itself (`swap()` and the like), `mempy()` and friends now check for identical input pointers for more efficiency. This could ideally just be a setting or something in the future, but right now this works.