VBOX FIXED: Make the stack bigger
We have done it! VBOX now boots, and is able to de-reference the PCI bus addresses. It seems the issue lied in a stack that was too small (I forgot to ever create a bigger stack in the kernel, and the prekernel one goes on to be used for the entire kernel).