Kernel: Rewrite `CMakeLists.txt`
Most things simply got formatting changes (I got lazy and just selected everything in Emacs `cmake-mode` and pressed `<tab>` to auto-indent everything).

The toolchain stuff should still be moved to it's own file, but that's not an issue right now.

Notable changes include:
- Use `ccache` if installed for all the speeds.
- System name and system version now explicitly specified.
- Variables that users may often change moved to the top of the file (`MACHINE`, `ARCH`)
`MACHINE` and `ARCH` are passed through to the kernel as pre-processor definitions, allowing for specific code to be ran on different hardware (or emulations there-of).
In the future, different header directories could be included based on arch and/or machine.