Toolchain: Windows Canadian Cross
Yes, it's correct; the LensorOS toolchain is officially running on Windows! More specifically, the `x86_64-w64-mingw32` host triplet. The toolchain binaries themselves must be built on Linux, but run on Windows, and generate executables for LensorOS, meaning this is a true Canadian Cross compiler (I always loved triple `C`s :^).

I think it's about time to provide pre-built toolchain binaries for both Windows and Linux; that would greatly simplify any onboarding for contributors, as well as make it easier to use the OS (without providing pre-built disk images; we're not quite that stable yet), but there are still some major issues with the system root configuration. The sysroot issues are temporarily fixed for freestanding programs (i.e. kernel only) using the `-I` command line flag; this will not work for userland programs due to system libraries not being present in the incorrect sysroot path.

As for solutions, it's possible the `--with-build-sysroot` flag could help, but documentation isn't clear on what it does; I guess it's time for me to start fiddling with the toolchain script that takes 45+ minutes to run each time... FUN!