Toolchain: Windows Oopsies
Too many conditional branches!!! I definitely want to patch if the directory exists, which it always should at this point in the script.

It's possible the patch has already been done; maybe a `patch_complete` file should be added to the end of the patches to be able to check if the patch is done or not? That's future stuff, anyway.