WinToolchain: Remove `BUILD_SYSROOT`
The documentation isn't clear, but it appears that `--with-build-sysroot` specifies the sysroot for the build system, not a sysroot to use while building (if that makes sense).

Either way, I'm going to revert back to the regular method (that actually compiles without fatal errors) and simply specify the `--sysroot` option in every use (a pain, but seemingly the only solution when doing this Canadian Cross thing).