WinToolchain: Set proper system root
At this point the LensorOS Toolchain works great on Windows, both freestanding and hosted (given the `--sysroot` option is used). A true Canadian Cross Compiler :^)

At this point, it's definitely the best idea to provide executables for both Windows and Linux... but I still have to update the documentation to include the new native Windows capabilities!