Toolchain: Make script (a little) smarter
At least you can sort of incrementally build parts of the toolchain by not deleting them if they succeed. Definitely saves a lot of time when toying with GCC sysroot stuff, as Binutils doesn't need to be re-built every time.

Not to mention the time saved in not having to download, extract, and patch the source archives every time.

Yes, this should have been part of the last PR but life goes on.