kstage1: Update `what function does` comment
I know it's best practice to just not have comments like these as they may become incorrect: the worst kind of comment. The reason I choose to keep it is two-fold. First, the list is a large help to me, since `kstage1` is hundred of lines long... Second, the list is a large help to newcomers to the codebase, as they can easily read what the code does in what order, then go figure out what the code looks like that performs that task. In the future, a more self-documenting approach would be better, but for now this is the solution.