VFS: Initial commit
The VFS, or Virtual File System, is a file system that exists only in memory. Its purpose is to manage other filesystems, allowing, say, a FAT formatted storage media to be accessed and modified in the same way that an ext2 formatted storage media could. This allows for a much better (and really only acceptable) user experience when dealing with files and such.

The basic idea is that the VFS is a filesystem that stores other filesystems and is able to communicate freely to, from, and between them. Each filesystem is "mounted" at a given path prefix; and can then be accessed by using a path of that prefix.

There's also the whole open file description thing. A file can not be read from or written to using a path; rather, an unsigned integer "File Descriptor" is used to find the entry in the system-wide list of open files. That entry ("Open File Description") stores the data necessary to read from/write to the file without having to do any calculations or figuring-out on where on the storage media the file resides. It's basically a simple file caching system.

This implementation is bound to be wrong, and buggy, and stupid; let's do our best to make it better.