Bootloader: Load kernel at 1MB physical
I have been trying to do this for forever!!! As it turns out, the issue lies in a missing linker flag that wasn't being passed to GCC: `-z max-page-size=0x1000`. This flag tells the linker to specify the page size as 4KiB. The default is the maximum possible page size on your architecture, for some odd reason (who needs 2MiB or even 1GiB pages?!). I'm not sure exactly how this was causing problems, but adding this flag solved the problems I was having.

Also, do checks that memory allocations don't fail (seems kind of important in a bootloader `:p`).