Kernel: `idtr storage` -> `idt storage`
The IDTR holds the offset (the base address) of the IDT; it doesn't need any storage itself.
The IDT on `x86_64` requires 4KiB of memory, so a place to store it is pre-allocated as a global variable (memory isn't yet initialized for allocation). This variable was renamed to better reflect it's actual purpose.