PCI: Fix major bugs
VirtualBox now agrees with my PCI enumeration! I was needlessly checking busses device headers when I should really only be checking devices and functions device headers.

Yes, now on both QEMU and VBOX a GPT partitioned disk (with a FAT32 partition) that is connected to an AHCI controller is able to be read from! It's very exciting to see it all coming together.