Bugfix: `VBOX` hangs when reading from AHCI port
Not sure exactly why it crashes, but I did find the exact line that tells me what's going wrong in `VBox.log`:

```
00:00:07.246217 EFI: VBoxDbg> loadimage64 '.efi' 0x1920000 LB 0x11000
00:00:07.449389 RTC: period=0x200 (512) 64 Hz
00:00:07.482343 AHCI#0P0: Read at offset 0 (0 bytes left) returned rc=VERR_NOT_SUPPORTED
00:00:11.471507 GUI: Request to close active machine-window.
00:00:11.474587 Changing the VM state from 'RUNNING' to 'SUSPENDING'
```

As you can see, it's an issue with VirtualBox itself not supporting reading at byte offset zero (although I imagine that may be a common case for hardware as well?). Or it's some other issue and this just happens to be right at the end of the logs when it hangs... Either way this is a simple quick fix for now.