Kernel: Add Windows `image_raw` target
`cmd` is able to properly parse the `mmd` commands with the path arguments that start with `/`, unlike Cygwin's `bash`. This means that the kernel can now be fully worked on from Windows without needing WSL for anything, from altering the code to running a VM (assuming pre-built binaries of LensorOS toolchain, LensorOS bootloader, and GNU mtools, hehe).