Create README2.md
This is a replacement for the original README that I've been changing and transforming to work for me over the past few months. This new README feels like a fresh start.

I do think it should include more images/visual appeal, but we'll get there. Rome in a day, and all.