Update README2.md
Add dependencies for building, as well as a link to VMWare Workstation Player; I was able to boot into LensorOS using it for the first time today. It isn't easy; it requires manually editing a file that doesn't *really* look like it should be, but there's no other way to access the EFI firmware. Other than that, it's kind of a breeze to set up, and I like how it works. I really need to set up scripts/build targets that automatically setup VBOX/VMWare now...