AHCI: Use cleaner API where possible
It's high time something is to be done about this AHCI mess. This is one of the first drivers I wrote and it clearly shows. I've went through and cleaned up the API, which actually makes even the low level read function nearly human readable. Also, none of the coding style fit the recommended coding style in `CONTRIBUTING.md`, so changes were made to be more consistent with the entire codebase; mainly member variables being PascalCase and not camelCase. There is still more to clean up and make better, as well as heaps more to implement, but this is a much better platform to jump off of, if you get my gist.