AHCI: Fix VBOX crash
I finally figured it out! I was setting the search flag on AHCI port storage devices with a type of `SATA` **or** `SATAPI`. This meant that CD-ROM devices were being treated exactly the same as hard disks, and this discrepancy caused issues that were fatal in VBOX (but not QEMU or VMWare). You could also say this was due to missing logic in the AHCI::PortController not checking for a proper PortType before continuing.

Either way, jobs done.