Add `image_iso` CMake target for Windows
Yep, still hate batch :^)

I wish it was easier in any language to get the current working directory!!! At least in bash it's sort of standardized with the BASH_SOURCE and `pwd` voodoo, but I feel like this isn't a feature that is that complicated, and it could really, really help a lot of scripts be built better.

I wonder if I could/should write `nu` scripts for any of these tasks?