Account for VMWare as a virtual machine
This reduces the overhead on the VM of trying to accomplish thousands of interrupts per second. Also, the HPET implementation works fine in VMWare as well as QEMU!

I wonder if it's something silly like the HPET isn't enabled by default in VirtualBox or something and I don't ever poll to see if it exists and exit if it doesn't... Or maybe I should be more aggressive about verifying the values found. I dunno, we'll get there eventually.