Update README.md
There was an issue in that the image would get generated before the build step **on some machines**. To fix this, simply use the `&&` operator that most shells have (`;` in nushell) which ensures that the build happens before any boot media generation.

I think this may be able to be fixed using better CMake dependencies, but it's not clear to me at the moment.