Kernel: Refactor String to allow conversion to c-style
Basically just keep a null terminator at the end no matter what, and
return a pointer to the base of the current string whenever `data()`
is called. `data_copy()` creates a new heap-allocated copy of the
null-terminated version of the string and returns the address of
that (free responsibility moves to caller even if string goes out of scope).