ELF: The loader, V0.0.1
This is by all means not the best way to commit things into a git
project, but I've worked for a few days on getting various things
working and I didn't have the patience while chasing down bugs to
commit everything as I went. Luckily, this can be treated as a
starting point and iterated upon from here.