VMM: Update
A parameter to alter page permissions has been added. This is
necessary for userland processes, as they must have the user bit set,
and the kernel must absolutely not.

This does create a breaking change, as the default mapping is present
but read-only. Most mapping in the entire kernel will need to have
their flags set explicitly.

Also, there is now optional degug output while mapping.