ELF: Don't triple fault
Something important must need to be globally mapped that isn't,
because when the identity mapping is removed from the page map,
it causes a triple fault upon switching to it.

Simple fix for now in just not unmapping the identity mapping, but a
better solution will definitely need to be found, soon.