Kernel: Nyan debug
Yep, it's as close to a rainbow as legacy ANSI/VT-100 color codes can
get :^)

I should figure out the true-color color codes and stuff and actually
make this a rainbow. It would be cool to have an alternating rainbow
in the foreground and background, as well, but might get a bit hard to read.