OH MY, IT WORKS
It actually works! I have spent months and months with exactly this
goal in my head, no idea what it would, could, or should look like,
and never knowing if I would actually make it... I did!

A userspace process is officially being run by LensorOS!!!

It's even integrated with the scheduler, and switches back and forth
between userland and kernel processes no problem.

Not only that, userland processes can make syscalls, *and it works*.

I can't believe everything I have been working on for so long is
finally a reality; not just an idea, but an implementation of one.

Today was especially bad, and this made me smile. Hopefully that will
keep me going for the next few years until I smile once more.

<3 <3 <3 <3 <3 <3 <3 <3 <3 <3 <3 <3 <3 <3 <3 <3 <3 <3 <3 <3 <3 <3 <3 <3