Syscalls use System V ABI, like ELF
As ELF is our executable format of choice, at least for now,
why not also follow the same calling convention?

I don't see any reason why this won't work.
*cut to a week later when I hate myself for thinking this*

The `syscall()` function in libc will be able to
load it's arguments into the proper registers,
as well as the return value(s), if any are even present.