Syscalls: `exit`
Currently this just gets the current PID then removes that
process from the list of running processes (`ProcessQueue`).

Also remove `ProcessState` as it is an idea that is no longer used.
I really like the `HALT AND CATCH FIRE` reference, so if this
ever comes back I would like to see this return as well :^).