Toolchain: Update to GCC 12.1.0
With this major version upgrade come some major changes!

The LensorOS GCC toolchain now supports auto-dereferencing of pointers
to structs and unions using the `.` operator. This means that the
de-referencing `->` operator is no longer necessary, but may still be
used explicitly. Yes, this puts the code-base "out of spec", but I
feel that this change is worth the deviation. All hail the `.`!