Renderer: Improve (eventual) concurrency
While we don't yet do multi-threading, it would be pretty impossible
to even do multi-processing using the old renderer, as it contained
a draw position as a member rather than a function argument. Now,
the renderer does not require caching and restoring of it's draw
position, which is a plus as well. All around an upgrade :^).