It works now, but only because the child process never gets run
This is a collection of a lot of the things I've tried that haven't
worked, as well as some minor comments and formatting.

The OS does boot now, without #PFs out the wazoo, but only because the
child process is started sleeping, and never exits that state. If it
is ever attempted to run the child process, things break
heavily. Obviously, this is non-ideal and we are working tirelessly to
find a solution.