¡It's Alive!
`fork` is now working as expected! It turns out it was the higher
levels of paging flags being overwritten by multiple map commands with
different permissions in a close area. With this fixed, we now have
the ability to duplicate the control flow of any program! How cool is
that?