¡Graphics in Userpace!
This disables the runtime loop in the kernel, as well as use of
gRend/gText after initial boot. Basically, the kernel now hands the ux
entirely off to userspace init process.

Some things to note:

1. The way of passing the framebuffer details from kernel to userspace
through `argv` string arguments is ... let's say this: *not ideal*.

2. It's probably a good idea to rename `stdout` to `init` and make
another process that simply writes to stdout