[Kernel/PIPE] Pipes are hard, mmkay
I've *began* to identify problems with pipes. It appears that `close`
is being called on the pipe driver itself each time `close` is called
on an fd associated with a pipe file. This isn't ideal, as we want the
pipe to stick around until *all* references to it are gone.