[Kernel/PIPE] `open` of same path returns same value
Before, we were having issues that when the file metadata was copied
via CopyUserspaceProcess in the scheduler, it was creating entirely
new pipe buffers in the new process (obviously not ideal). This solves
that, and now only one pipe buffer is created and deleted in the span
of the one call to `pipe` and the four calls to `close`, two each on
each side of the `fork`.