[Kernel] Free unused page maps from separate process
The kernel process now has a purpose for sticking around! It can free
the physical memory used to represent the process' virtual memory
mappings. For some reason, it's a bit scuffed, and there are a lot of
garbage values in each page table somehow (I'll have to figure that
one out). It may be worth it to start the init process with a fresh
page map instead of cloning the kernel's.