[Build] Use new CMake script to make raw boot media images
This shortens/simplifies the main cmakelists, which is always
nice. If we could also do this for gpt and iso boot media, while
keeping dependencies intact, that would be really cool.

Also, make `image_gpt` the default, when it is present (as it is the
assumed default of LensorOS kernel/userspace, at the moment).