[Build] ¡Userspace Build Process Overhaul!
While it still isn't *absolutely* ideal, I've reduced the number of
places that need updated when a new userspace library/program is added
to just two: `scripts/mkfat32lensorosdata.cmake` and `user/CMakeLists.txt`.

Something to note: the pre-built userspace programs are no longer
included in the repository. We could always track `user/bin`, which is
where they are now installed to and from. However, I don't think this
is necessary: when building a release for LensorOS, it would make
sense to include these, but not in the develoment repository, I don't
think. If it causes issues, we can, as I said, just track `user/bin`.