[Kernel/Pipe] Refactor `PipeDriver`
The PipeDriver was implemented in a limited way: a pipe only created a
single `std::shared_ptr<FileMetadata>`. This is an issue because that
means we only knew when both ends of the pipe had closed completely,
but not one or the other. This fixes that by altering `lay_pipe` to
vend two shared_ptrs, one each for the read and write end of the
pipe. This should hopefully make it possible to implement a sort-of
blocking-on-read functionality, where a return value of `-2` or
something signifies "try again", and we sleep the process until we
know the try will succeed (aka data was written to the pipe).

Also in this commit: File descriptors are now copied upon `fork()`
differently. Before, we relied on an invariant "opening the same path
in a given driver will return the same FileMetadata". However this
isn't necessarily the case. So it's better to just create a new
shared_ptr to the same FileMetadata object and add that (I think).