[Kernel] Fix problem with SystemDevice iterator invalidation in `kstage1`
Basically, because we are appending to a container as we are looping
over it, we need to ensure iterators don't get invalidated. To do
this, we re-calculate the iterator every iteration of the loop using
an index. I'm pretty sure we were only getting away with this before
due to the SYSTEM->Devices vector not growing large enough to resize
in one of these loops. With the addition of the e1000 device, however,
this has apparently changed.