[Kernel/E1000] Enough documentation to knock your socks off
This goes through and documents just the very basic details of all of
the registers available in the address space given through BAR0 or
BAR1:BAR0.

For a lot of these registers, it would be beneficial still to actually
give bit-layout descriptions, as done with `REG_EECD` and `REG_EEPROM`.