[SYS$] Abstract "blocking" functionality to the VFS
Before, this was in each driver; this way, we can have it so that all
the driver has to do to block the process is return `-2`.

Eventually, we should even move the "PIDsWaiting" vector into the
FileMetadata, and maybe even change the type it's storing; instead of
just a `pid_t`, we could store a pid_t + a reason for waiting; some
processes could block until *anything* happens to a file, some just
until a write, some just until a read, etc.