[Kernel/Memory] Abstract cloning of page map
Honestly we should probably have like a flag filter mask that we use
but for now this will do. Basically, we eventually want to, upon
`fork`ing, copy the page map but actually copy every entry as
readonly. Then, we "simply" don't allocate new physical memory for any
of the process' virtual pages; after all, it's an exact copy of a
program that is already loaded. Why load it twice? Then, we need to
have a page fault caused by a write from userspace to a valid memory
region actually do the copy. This is called copy-on-write, and will
help efficiency dramatically once implemented, as `libc` .text section
will only ever be loaded into physical memory once, etc etc.